+++ src/core/lotus_entitycomponent.hpp
@@ -8,7 +8,7 @@ class EntityComponent
 protected:
 	Entity *m_entity;
 public:
-	virtual void update(double delta) {}
+	virtual void update(float delta) {}
 	virtual void render(IRenderer *renderer) const {}
 
 	inline Transform &getTransform() const { return m_entity->getTransform(); }
